chore(broker): remove unused code and revert partitionIds to int
related to #414

+++ broker-core/src/main/java/io/zeebe/broker/clustering/handler/TopicLeader.java
@@ -73,7 +73,7 @@ public class TopicLeader extends UnpackedObject
         return this;
     }
 
-    public long getPartitionId()
+    public int getPartitionId()
     {
         return partitionIdProp.getValue();
     }
